Separate CMS & CSI module files; update README
